+++ modules/snooker/src/test/ts/browser/lookup/BlocksTest.ts
@@ -12,7 +12,7 @@ UnitTest.test('BlocksTest', () => {
     return elem;
   };
   const s = (elemText: string, rowspan: number, colspan: number) => Structs.detail(createCell(elemText), rowspan, colspan);
-  const f = (cells: Structs.Detail[], section: 'tbody' | 'thead' | 'tfoot') => Structs.rowdata(SugarElement.fromTag('tr'), cells, section);
+  const f = (cells: Structs.Detail[], section: 'tbody' | 'thead' | 'tfoot') => Structs.rowdetail(SugarElement.fromTag('tr'), cells, section);
 
   const warehouse = Warehouse.generate([
     f([ s('a', 1, 1), s('b', 1, 2) ], 'thead'),