+++ modules/snooker/src/test/ts/browser/operate/ModificationOperationsTest.ts
@@ -10,6 +10,7 @@ import BrowserTestGenerator from 'ephox/snooker/test/BrowserTestGenerator';
 
 UnitTest.test('ModificationOperationsTest', () => {
   const r = Structs.rowcells;
+  const re = () => SugarElement.fromTag('tr');
   const en = (content: string, isNew: boolean) => {
     const elem = SugarElement.fromTag('td');
     Html.set(elem, content);
@@ -17,18 +18,21 @@ UnitTest.test('ModificationOperationsTest', () => {
   };
   const mapToStructGrid = (grid: Structs.ElementNew[][]) => {
     return Arr.map(grid, (row) => {
-      return Structs.rowcells(row, 'tbody');
+      return Structs.rowcells(re(), row, 'tbody', false);
     });
   };
 
-  const assertGrids = (expected: Structs.RowCells[], actual: Structs.RowCells[]) => {
+  const assertGrids = (expected: Structs.RowCells[], actual: Structs.RowCells[], checkIsNew: boolean) => {
     assert.eq(expected.length, actual.length);
     Arr.each(expected, (row, i) => {
       Arr.each(row.cells, (cell, j) => {
         Assertions.assertHtml('Expected elements to have the same HTML', Html.getOuter(cell.element), Html.getOuter(actual[i].cells[j].element));
         assert.eq(cell.isNew, actual[i].cells[j].isNew);
       });
-      assert.eq(row.section, actual[i].section);
+      assert.eq(row.section, actual[i].section, 'section type');
+      if (checkIsNew) {
+        assert.eq(row.isNew, actual[i].isNew, 'is new row');
+      }
     });
   };
 
@@ -36,15 +40,15 @@ UnitTest.test('ModificationOperationsTest', () => {
 
   // Test basic insert column
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], example: number, index: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], example: number, index: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.insertColumnAt(grid, index, example, compare, Generators.modification(BrowserTestGenerator()).getOrInit);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], example: number, index: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, example, index);
+      check(structExpected, structGrid, example, index, false);
     };
 
     checkBody([], [], 0, 0);
@@ -96,49 +100,49 @@ UnitTest.test('ModificationOperationsTest', () => {
 
     check(
       [
-        r([ en('a', false), en('?_0', true) ], 'thead'),
-        r([ en('b', false), en('?_1', true) ], 'tbody')
+        r(re(), [ en('a', false), en('?_0', true) ], 'thead', false),
+        r(re(), [ en('b', false), en('?_1', true) ], 'tbody', false)
       ],
       [
-        r([ en('a', false) ], 'thead'),
-        r([ en('b', false) ], 'tbody')
+        r(re(), [ en('a', false) ], 'thead', false),
+        r(re(), [ en('b', false) ], 'tbody', false)
       ], 0, 1
     );
 
     check(
       [
-        r([ en('?_0', true), en('a', false) ], 'thead'),
-        r([ en('?_1', true), en('b', false) ], 'tbody')
+        r(re(), [ en('?_0', true), en('a', false) ], 'thead', false),
+        r(re(), [ en('?_1', true), en('b', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false) ], 'thead'),
-        r([ en('b', false) ], 'tbody')
+        r(re(), [ en('a', false) ], 'thead', false),
+        r(re(), [ en('b', false) ], 'tbody', false)
       ], 0, 0
     );
 
     check(
       [
-        r([ en('a', false), en('a', false), en('a', false) ], 'thead'),
-        r([ en('b', false), en('?_0', true), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('a', false), en('a', false) ], 'thead', false),
+        r(re(), [ en('b', false), en('?_0', true), en('c', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('a', false) ], 'thead'),
-        r([ en('b', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('a', false) ], 'thead', false),
+        r(re(), [ en('b', false), en('c', false) ], 'tbody', false)
       ], 0, 1
     );
   })();
 
   // Test basic insert row
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], example: number, index: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], example: number, index: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.insertRowAt(grid, index, example, compare, Generators.modification(BrowserTestGenerator()).getOrInit);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], example: number, index: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, example, index);
+      check(structExpected, structGrid, example, index, false);
     };
 
     checkBody([[ en('?_0', true) ], [ en('a', false) ]], [[ en('a', false) ]], 0, 0);
@@ -159,38 +163,38 @@ UnitTest.test('ModificationOperationsTest', () => {
 
     check(
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('?_0', true), en('?_1', true), en('?_2', true) ], 'thead'),
-        r([ en('d', false), en('e', false), en('f', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('?_0', true), en('?_1', true), en('?_2', true) ], 'thead', true),
+        r(re(), [ en('d', false), en('e', false), en('f', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('d', false), en('e', false), en('f', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('d', false), en('e', false), en('f', false) ], 'tbody', false)
       ], 0, 1);
 
     check(
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('?_0', true), en('?_1', true), en('?_2', true) ], 'tbody'),
-        r([ en('d', false), en('e', false), en('f', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('?_0', true), en('?_1', true), en('?_2', true) ], 'tbody', true),
+        r(re(), [ en('d', false), en('e', false), en('f', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('d', false), en('e', false), en('f', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('d', false), en('e', false), en('f', false) ], 'tbody', false)
       ], 1, 1);
   })();
 
   // Test basic delete column
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], index: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], index: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.deleteColumnsAt(grid, [ index ]);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], index: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, index);
+      check(structExpected, structGrid, index, false);
     };
 
     checkBody([], [[ en('a', false) ]], 0);
@@ -206,26 +210,26 @@ UnitTest.test('ModificationOperationsTest', () => {
       ], 1);
     check(
       [
-        r([ en('a', false), en('b', false) ], 'thead'),
-        r([ en('c', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('c', false), en('c', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('b', false) ], 'thead'),
-        r([ en('c', false), en('c', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('c', false), en('c', false), en('c', false) ], 'tbody', false)
       ], 1);
   })();
 
   // Test delete multiple columns
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], indexes: number[]) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], indexes: number[], checkIsNew: boolean = true) => {
       const actual = ModificationOperations.deleteColumnsAt(grid, indexes);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], indexes: number[]) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, indexes);
+      check(structExpected, structGrid, indexes, false);
     };
 
     checkBody([], [[ en('a', false), en('b', false) ]], [ 0, 1 ]);
@@ -241,26 +245,26 @@ UnitTest.test('ModificationOperationsTest', () => {
       ], [ 1 ]);
     check(
       [
-        r([ en('a', false) ], 'thead'),
-        r([ en('c', false) ], 'tbody')
+        r(re(), [ en('a', false) ], 'thead', false),
+        r(re(), [ en('c', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('c', false), en('c', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('c', false), en('c', false), en('c', false) ], 'tbody', false)
       ], [ 1, 2 ]);
   })();
 
   // Test basic delete row
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], index: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], index: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.deleteRowsAt(grid, index, index);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], index: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, index);
+      check(structExpected, structGrid, index, false);
     };
 
     checkBody([], [[ en('a', false) ]], 0);
@@ -278,26 +282,26 @@ UnitTest.test('ModificationOperationsTest', () => {
 
     check(
       [
-        r([ en('a', false), en('b', false), en('b', false) ], 'thead'),
-        r([ en('c', false), en('c', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('c', false), en('c', false), en('c', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('b', false) ], 'thead'),
-        r([ en('a', false), en('b', false), en('b', false) ], 'tbody'),
-        r([ en('c', false), en('c', false), en('c', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('a', false), en('b', false), en('b', false) ], 'tbody', false),
+        r(re(), [ en('c', false), en('c', false), en('c', false) ], 'tbody', false)
       ], 1);
   })();
 
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], exRow: number, exCol: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], exRow: number, exCol: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.splitCellIntoColumns(grid, exRow, exCol, Fun.tripleEquals, Generators.modification(BrowserTestGenerator()).getOrInit);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], exRow: number, exCol: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, exRow, exCol);
+      check(structExpected, structGrid, exRow, exCol, false);
     };
 
     // splitting simple tables without existing colspans
@@ -460,41 +464,41 @@ UnitTest.test('ModificationOperationsTest', () => {
     );
     check(
       [
-        r([ en('a', false), en('b', false), en('c', false), en('?_0', true) ], 'thead'),
-        r([ en('d', false), en('e', false), en('e', false), en('e', false) ], 'tbody'),
-        r([ en('g', false), en('g', false), en('i', false), en('i', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false), en('?_0', true) ], 'thead', false),
+        r(re(), [ en('d', false), en('e', false), en('e', false), en('e', false) ], 'tbody', false),
+        r(re(), [ en('g', false), en('g', false), en('i', false), en('i', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false), en('c', false) ], 'thead'),
-        r([ en('d', false), en('e', false), en('e', false) ], 'tbody'),
-        r([ en('g', false), en('g', false), en('i', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('d', false), en('e', false), en('e', false) ], 'tbody', false),
+        r(re(), [ en('g', false), en('g', false), en('i', false) ], 'tbody', false)
       ], 0, 2
     );
 
     check(
       [
-        r([ en('a', false), en('a', false), en('a', false), en('c', false) ], 'thead'),
-        r([ en('a', false), en('a', false), en('a', false), en('a', false) ], 'tbody'),
-        r([ en('g', false), en('?_0', true), en('h', false), en('i', false) ], 'tbody')
+        r(re(), [ en('a', false), en('a', false), en('a', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('a', false), en('a', false), en('a', false), en('a', false) ], 'tbody', false),
+        r(re(), [ en('g', false), en('?_0', true), en('h', false), en('i', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('a', false), en('c', false) ], 'thead'),
-        r([ en('a', false), en('a', false), en('a', false) ], 'tbody'),
-        r([ en('g', false), en('h', false), en('i', false) ], 'tbody')
+        r(re(), [ en('a', false), en('a', false), en('c', false) ], 'thead', false),
+        r(re(), [ en('a', false), en('a', false), en('a', false) ], 'tbody', false),
+        r(re(), [ en('g', false), en('h', false), en('i', false) ], 'tbody', false)
       ], 2, 0
     );
   })();
 
   (() => {
-    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], exRow: number, exCol: number) => {
+    const check = (expected: Structs.RowCells[], grid: Structs.RowCells[], exRow: number, exCol: number, checkIsNew: boolean = true) => {
       const actual = ModificationOperations.splitCellIntoRows(grid, exRow, exCol, Fun.tripleEquals, Generators.modification(BrowserTestGenerator()).getOrInit);
-      assertGrids(expected, actual);
+      assertGrids(expected, actual, checkIsNew);
     };
 
     const checkBody = (expected: Structs.ElementNew[][], grid: Structs.ElementNew[][], exRow: number, exCol: number) => {
       const structExpected = mapToStructGrid(expected);
       const structGrid = mapToStructGrid(grid);
-      check(structExpected, structGrid, exRow, exCol);
+      check(structExpected, structGrid, exRow, exCol, false);
     };
 
     // splitting simple tables without existing rowspans
@@ -674,13 +678,13 @@ UnitTest.test('ModificationOperationsTest', () => {
 
     check(
       [
-        r([ en('a', false), en('b', false) ], 'thead'),
-        r([ en('a', false), en('?_0', true) ], 'thead'),
-        r([ en('c', false), en('d', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('a', false), en('?_0', true) ], 'thead', true),
+        r(re(), [ en('c', false), en('d', false) ], 'tbody', false)
       ],
       [
-        r([ en('a', false), en('b', false) ], 'thead'),
-        r([ en('c', false), en('d', false) ], 'tbody')
+        r(re(), [ en('a', false), en('b', false) ], 'thead', false),
+        r(re(), [ en('c', false), en('d', false) ], 'tbody', false)
       ], 0, 1
     );
   })();