+++ modules/tinymce/src/plugins/table/main/ts/core/TableFormats.ts
@@ -6,44 +6,47 @@
  */
 
 import Editor from 'tinymce/core/api/Editor';
+const genericBase = {
+  remove_similar: true,
+  inherit: false
+};
+
+const cellBase = {
+  selector: 'td,th',
+  ...genericBase
+};
 
 const cellFormats = {
   tablecellbackgroundcolor: {
-    selector: 'td,th',
     styles: { backgroundColor: '%value' },
-    remove_similar: true
+    ...cellBase
   },
   tablecellverticalalign: {
-    selector: 'td,th',
     styles: {
       'vertical-align': '%value'
     },
-    remove_similar: true
+    ...cellBase
   },
   tablecellbordercolor: {
-    selector: 'td,th',
     styles: { borderColor: '%value' },
-    remove_similar: true
+    ...cellBase
   },
   tablecellclass: {
-    selector: 'td,th',
     classes: [ '%value' ],
-    remove_similar: true
+    ...cellBase
   },
   tableclass: {
     selector: 'table',
     classes: [ '%value' ],
-    remove_similar: true
+    ...genericBase
   },
   tablecellborderstyle: {
-    selector: 'td,th',
     styles: { borderStyle: '%value' },
-    remove_similar: true
+    ...cellBase
   },
   tablecellborderwidth: {
-    selector: 'td,th',
     styles: { borderWidth: '%value' },
-    remove_similar: true
+    ...cellBase
   }
 };