+++ modules/tinymce/src/plugins/table/main/ts/ui/CellDialog.ts
@@ -16,7 +16,6 @@ import { Dialog } from 'tinymce/core/api/ui/Ui';
 import * as Styles from '../actions/Styles';
 import * as Events from '../api/Events';
 import { hasAdvancedCellTab } from '../api/Settings';
-import { switchCellType } from '../core/TableSections';
 import * as Util from '../core/Util';
 import * as TableSelection from '../selection/TableSelection';
 import * as CellDialogGeneralTab from './CellDialogGeneralTab';
@@ -30,23 +29,22 @@ interface SelectedCell {
   column: Optional<HTMLTableColElement>;
 }
 
-const getSelectedCells = (cells: SugarElement<HTMLTableCellElement>[]): Optional<SelectedCell[]> =>
-  TableLookup.table(cells[0]).map((table) => {
-    const warehouse = Warehouse.fromTable(table);
+const getSelectedCells = (table: SugarElement<HTMLTableElement>, cells: SugarElement<HTMLTableCellElement>[]): SelectedCell[] => {
+  const warehouse = Warehouse.fromTable(table);
 
-    const allCells = Warehouse.justCells(warehouse);
+  const allCells = Warehouse.justCells(warehouse);
 
-    const filtered = Arr.filter(allCells, (cellA) =>
-      Arr.exists(cells, (cellB) =>
-        Compare.eq(cellA.element, cellB)
-      )
-    );
+  const filtered = Arr.filter(allCells, (cellA) =>
+    Arr.exists(cells, (cellB) =>
+      Compare.eq(cellA.element, cellB)
+    )
+  );
 
-    return Arr.map(filtered, (cell) => ({
-      element: cell.element.dom,
-      column: Warehouse.getColumnAt(warehouse, cell.column).map((col) => col.element.dom)
-    }));
-  });
+  return Arr.map(filtered, (cell) => ({
+    element: cell.element.dom,
+    column: Warehouse.getColumnAt(warehouse, cell.column).map((col) => col.element.dom)
+  }));
+};
 
 const updateSimpleProps = (modifier: DomModifier, colModifier: DomModifier, data: CellData) => {
   modifier.setAttrib('scope', data.scope);
@@ -62,73 +60,86 @@ const updateAdvancedProps = (modifier: DomModifier, data: CellData) => {
   modifier.setFormat('tablecellborderwidth', Util.addPxSuffix(data.borderwidth));
 };
 
-// NOTES:
+/*
+  NOTES:
 
-// When applying to a single cell, values can be falsy. That is
-// because there should be a consistent value across the cell
-// selection, so it should also be possible to toggle things off.
+  When applying to a single cell, values can be falsy. That is
+  because there should be a consistent value across the cell
+  selection, so it should also be possible to toggle things off.
 
-// When applying to multiple cells, values must be truthy to be set.
-// This is because multiple cells might have different values, and you
-// don't want a blank value to wipe out their original values. Note,
-// how as part of this, it doesn't remove any original alignment before
-// applying any specified alignment.
+  When applying to multiple cells, values must be truthy to be set.
+  This is because multiple cells might have different values, and you
+  don't want a blank value to wipe out their original values. Note,
+  how as part of this, it doesn't remove any original alignment before
+  applying any specified alignment.
+ */
+const applyStyleData = (editor: Editor, cells: SelectedCell[], data: CellData) => {
+  const isSingleCell = cells.length === 1;
+  Arr.each(cells, (item) => {
+    const cellElm = item.element;
+    const modifier = isSingleCell ? DomModifier.normal(editor, cellElm) : DomModifier.ifTruthy(editor, cellElm);
+    const colModifier = item.column.map((col) =>
+      isSingleCell ? DomModifier.normal(editor, col) : DomModifier.ifTruthy(editor, col)
+    ).getOr(modifier);
+
+    updateSimpleProps(modifier, colModifier, data);
+
+    if (hasAdvancedCellTab(editor)) {
+      updateAdvancedProps(modifier, data);
+    }
+
+    // Remove alignment
+    if (isSingleCell) {
+      Styles.unApplyAlign(editor, cellElm);
+      Styles.unApplyVAlign(editor, cellElm);
+    }
+
+    // Apply alignment
+    if (data.halign) {
+      Styles.applyAlign(editor, cellElm, data.halign);
+    }
+
+    // Apply vertical alignment
+    if (data.valign) {
+      Styles.applyVAlign(editor, cellElm, data.valign);
+    }
+  });
+};
+
+const applyStructureData = (editor: Editor, data: CellData) => {
+  // Switch cell type if applicable. Note that we specifically tell the command to not fire events
+  // as we'll batch the events and fire a `TableModified` event at the end of the updates.
+  editor.execCommand('mceTableCellType', false, { type: data.celltype, no_events: true });
+};
 
 const applyCellData = (editor: Editor, cells: SugarElement<HTMLTableCellElement>[], oldData: CellData, data: CellData) => {
-  const isSingleCell = cells.length === 1;
-
   const modifiedData = Obj.filter(data, (value, key) => oldData[key] !== value);
 
   if (Obj.size(modifiedData) > 0 && cells.length >= 1) {
-    /*
-      Retrieve the table before the cells are modified
-      as there is a case where cells are replaced and
-      the reference will be lost when trying to fire events.
-    */
-    const tableOpt = TableLookup.table(cells[0]);
+    // Retrieve the table before the cells are modified as there is a case where cells
+    // are replaced and the reference will be lost when trying to fire events.
+    TableLookup.table(cells[0]).each((table) => {
+      const selectedCells = getSelectedCells(table, cells);
 
-    getSelectedCells(cells).each((selectedCells) => {
-      Arr.each(selectedCells, (item) => {
-        // Switch cell type if applicable
-        const cellElm = switchCellType(editor, item.element, data.celltype);
-        const modifier = isSingleCell ? DomModifier.normal(editor, cellElm) : DomModifier.ifTruthy(editor, cellElm);
-        const colModifier = item.column.map((col) =>
-          isSingleCell ? DomModifier.normal(editor, col) : DomModifier.ifTruthy(editor, col)
-        ).getOr(modifier);
+      // style modified if there's at least one other change apart from 'celltype' and 'scope'
+      const styleModified = Obj.size(Obj.filter(modifiedData, (_value, key) => key !== 'scope' && key !== 'celltype')) > 0;
+      const structureModified = Obj.has(modifiedData, 'celltype');
 
-        updateSimpleProps(modifier, colModifier, data);
+      // Update the cells styling using the dialog data
+      if (styleModified || Obj.has(modifiedData, 'scope')) {
+        applyStyleData(editor, selectedCells, data);
+      }
 
-        if (hasAdvancedCellTab(editor)) {
-          updateAdvancedProps(modifier, data);
-        }
+      // Update the cells structure using the dialog data
+      if (structureModified) {
+        applyStructureData(editor, data);
+      }
 
-        // Remove alignment
-        if (isSingleCell) {
-          Styles.unApplyAlign(editor, cellElm);
-          Styles.unApplyVAlign(editor, cellElm);
-        }
-
-        // Apply alignment
-        if (data.halign) {
-          Styles.applyAlign(editor, cellElm, data.halign);
-        }
-
-        // Apply vertical alignment
-        if (data.valign) {
-          Styles.applyVAlign(editor, cellElm, data.valign);
-        }
+      Events.fireTableModified(editor, table.dom, {
+        structure: structureModified,
+        style: styleModified,
       });
     });
-
-    // style modified if there's at least one other change apart from 'celltype' and 'scope'
-    const styleModified = Obj.size(Obj.filter(modifiedData, (_value, key) => key !== 'scope' && key !== 'celltype')) > 0;
-
-    tableOpt.each(
-      (table) => Events.fireTableModified(editor, table.dom, {
-        structure: Obj.has(modifiedData, 'celltype'),
-        style: styleModified,
-      })
-    );
   }
 };
 
@@ -143,8 +154,8 @@ const onSubmitCellForm = (editor: Editor, cells: SugarElement<HTMLTableCellEleme
 };
 
 const getData = (editor: Editor, cells: SugarElement<HTMLTableCellElement>[]) => {
-  const cellsData = getSelectedCells(cells).map((selectedCells) =>
-    Arr.map(selectedCells, (item) =>
+  const cellsData = TableLookup.table(cells[0]).map((table) =>
+    Arr.map(getSelectedCells(table, cells), (item) =>
       Helpers.extractDataFromCellElement(editor, item.element, hasAdvancedCellTab(editor), item.column)
     )
   );